refactor(realm): extract websocket observer from js-realm-shared
