+++ packages/webapp/src/kernel/realm/js-realm-shared.ts
@@ -43,15 +43,13 @@ import {
   NodeExitError,
 } from './realm-node-shims.js';
 import { type RealmPortLike, RealmRpcClient } from './realm-rpc.js';
+import { createWsObserverApi } from './realm-ws-observer.js';
 import type {
   RealmDoneMsg,
   RealmInitMsg,
   RealmRpcChannel,
   SerializedFetchResponse,
   TabHandle,
-  WsSelector,
-  WsSink,
-  WsSubscriberInfo,
 } from './realm-types.js';
 import { createSkillGlobal, type SkillFsBridge } from './skill-global.js';
 import { SyncFsCache, type SyncFsSnapshot } from './sync-fs-cache.js';
@@ -431,7 +429,7 @@ function createBrowserBridge(rpc: RealmRpcClient) {
   };
 }
 
-function resolveTargetId(tab: TabHandle | string): string {
+export function resolveTargetId(tab: TabHandle | string): string {
   if (typeof tab === 'string') return tab;
   if (tab && typeof tab === 'object' && typeof tab.targetId === 'string') return tab.targetId;
   throw new TypeError('browser: expected a tab handle or targetId string');
@@ -784,98 +782,6 @@ function normalizeMatchUrl(options: { matchUrl?: string | RegExp }): { matchUrl?
   if (options.matchUrl === undefined) return {};
   return {
     matchUrl: options.matchUrl instanceof RegExp ? options.matchUrl.source : options.matchUrl,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// `browser.websocket` — declarative WebSocket observer
-// ---------------------------------------------------------------------------
-
-/**
- * Builder for a `browser.websocket.on(tab, opts)` chain. The selector
- * (`.filter`) and sink (`.forward`) are collected on the builder; the
- * actual subscriber is created by the await on `.forward(...)`, which
- * resolves to a {@link WsSubscriberHandle}.
- */
-interface WsObserverBuilder {
-  filter(selector: WsSelector): WsObserverBuilder;
-  forward(sink: WsSink): Promise<WsSubscriberHandle>;
-}
-
-interface WsSubscriberHandle extends WsSubscriberInfo {
-  update(patch: {
-    urlMatch?: string | RegExp | null;
-    filter?: WsSelector | null;
-  }): Promise<WsSubscriberInfo>;
-  close(): Promise<boolean>;
-}
-
-interface WsObserverApi {
-  on(tab: TabHandle | string, opts?: { urlMatch?: string | RegExp }): WsObserverBuilder;
-  list(): Promise<WsSubscriberInfo[]>;
-}
-
-/**
- * Construct the realm-side `browser.websocket` chainable API. All
- * actual work happens host-side; this file just shapes the builder
- * surface and forwards JSON-safe payloads over the `browser` RPC
- * channel.
- */
-function createWsObserverApi(rpc: RealmRpcClient): WsObserverApi {
-  function makeHandle(info: WsSubscriberInfo): WsSubscriberHandle {
-    return {
-      ...info,
-      async update(patch): Promise<WsSubscriberInfo> {
-        const wire: { urlMatch?: string | null; filter?: WsSelector | null } = {};
-        if (patch.urlMatch !== undefined) {
-          wire.urlMatch =
-            patch.urlMatch === null
-              ? null
-              : patch.urlMatch instanceof RegExp
-                ? patch.urlMatch.source
-                : patch.urlMatch;
-        }
-        if (patch.filter !== undefined) wire.filter = patch.filter;
-        return rpc.call<WsSubscriberInfo>('browser', 'wsUpdate', [info.id, wire]);
-      },
-      async close(): Promise<boolean> {
-        return rpc.call<boolean>('browser', 'wsClose', [info.id]);
-      },
-    };
-  }
-
-  return {
-    on(tab, opts = {}) {
-      const targetId = resolveTargetId(tab);
-      const urlMatch =
-        opts.urlMatch === undefined
-          ? undefined
-          : opts.urlMatch instanceof RegExp
-            ? opts.urlMatch.source
-            : opts.urlMatch;
-      let selector: WsSelector | undefined;
-      const builder: WsObserverBuilder = {
-        filter(next) {
-          if (typeof next === 'function' || typeof next === 'string') {
-            throw new TypeError(
-              'browser.websocket: filter must be a declarative JSON object, not a function or string'
-            );
-          }
-          selector = next;
-          return builder;
-        },
-        async forward(sink) {
-          const info = await rpc.call<WsSubscriberInfo>('browser', 'wsObserve', [
-            { targetId, urlMatch, filter: selector, forward: sink },
-          ]);
-          return makeHandle(info);
-        },
-      };
-      return builder;
-    },
-    async list() {
-      return rpc.call<WsSubscriberInfo[]>('browser', 'wsList', []);
-    },
   };
 }
